requirements.txt
scrapling[fetchers]
python-dotenv
schedule
requests
